Add allowDefaults hook to TouchAction.prototype
- This allows the application to override the default behavior of
  TouchAction, which is to prevent the default browser action for
gestures that are not registered to be recognized.  This was to address
an issue in our application, where we wanted to pan child elements
horizontally, but still allow a vertical pan gesture to bubble to the
parent element, which is scrollable.  There were other ways to address
this, but this was the simplest.